% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clim_ar5.R
\name{clim_dist_monthly}
\alias{clim_dist_monthly}
\title{Extract climate data and compute distributions}
\usage{
clim_dist_monthly(model_idx, cells, inputs, in_dir = snapdef()$ar5dir,
  out_dir = snapdef()$ar5dir_dist_monthly, na.rm = TRUE,
  density.args = list(n = 200, adjust = 0.1), sample.size = 10000,
  verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{model_idx}{iterator, model index in list of climate models/directories.}

\item{cells}{data frame of grid cells in raster maps corresponding to various standard SNAP spatial polygons,
e.g., the table from \code{snapdef()$akcan1km2km}.}

\item{inputs}{data frame of inputs. See \code{clim_inputs_table}.}

\item{in_dir}{input directory, e.g., \code{snapdef()$ar5dir}.}

\item{out_dir}{output directory, e.g., \code{snapdef()$ar5dir_dist_monthly}}

\item{na.rm}{logical, remove NAs.}

\item{density.args}{arguments list passed to \code{density}.}

\item{sample.size}{numeric, sample size.}

\item{verbose}{logical, verbose progress.}

\item{overwrite}{logical, overwrite existing files.}
}
\value{
invisible, writes files.
}
\description{
Extract climate data and estimate monthly spatial probability distributions.
}
\examples{
\dontrun{mclapply(1:nrow(inputs), get_data, cells = cells,
  inputs = inputs, in_dir = getwd(), out_dir = out_dir, mc.cores = 2)}
}
