% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clim_ar5.R
\name{clim_dist_seasonal}
\alias{clim_dist_seasonal}
\title{Compute seasonal climate data spatial distributions}
\usage{
clim_dist_seasonal(i, files, out_dir = snapdef()$ar5dir_dist_seasonal,
  density.args = list(n = 200, adjust = 0.1))
}
\arguments{
\item{i}{iterator for files.}

\item{files}{vector of input files. See details.}

\item{out_dir}{output directory.}

\item{density.args}{arguments list passed to \code{density}.}
}
\value{
invisible, writes files.
}
\description{
Compute seasonal climate data spatial probability distributions.
}
\details{
Seasons are DJF, MAM, JJA and SON 3-month averages. A fifth "season" of full annual averages is also included.
For efficiency, this function operates on outputs from \code{clim_dist_monthly}. It does not need to redundantly
access source downscaled geotiffs.

\code{files} typically come from \code{snapdef()$ar5dir_dist_monthly}. See example for how to list files.
}
\examples{
\dontrun{
in_dir <- snapdef()$ar5dir_dist_monthly
files <- list.files(in_dir, pattern = ".rds$", full.names = TRUE, recursive = TRUE)
parallel::mclapply(seq_along(files), clim_dist_seasonal, files = files, mc.cores = 32)
 }
}
