% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clim_ar5.R
\name{clim_locs_prep}
\alias{clim_locs_prep}
\title{Extract climate values for point locations}
\usage{
clim_locs_prep(inputs, in_dir = snapdef()$ar5dir,
  out_dir = snapdef()$ar5dir_locs_prep, verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{inputs}{data frame of inputs (one row). See details.}

\item{in_dir}{input directory, e.g., \code{snapdef()$ar5dir}.}

\item{out_dir}{output directory, e.g., \code{snapdef()$ar5dir_locs_prep}}

\item{verbose}{logical, verbose progress.}

\item{overwrite}{logical, overwrite existing files.}
}
\description{
Extract monthly climate values for point locations by RCP, climate model and climate variable.
}
\details{
\code{inputs} generally comes from \link{clim_inputs_table}. \code{clim_locs_prep} processes data sets referred to by
one row of this data frame at a time. This function generates intermediary temp files used susbsequently by \code{\link{clim_locs}}
and can be deleted afterward if not needed for other uses where it is convenient to have extracted point location climate data
segmented into files by RCP, model and variable as opposed to by point location.
}
\examples{
\dontrun{
library(dplyr)
library(parallel)
inputs <- clim_inputs_table() \%>\%
  filter(!(model == "ts40" & var \%in\% c("tasmin", "tasmax")))
mclapply(split(inputs, 1:nrow(inputs)), clim_locs_prep, mc.cores = 32)
}
}
