% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_cells.R
\name{snap_poly_list}
\alias{snap_poly_list}
\title{Generate nested list of spatial polygons data frames}
\usage{
snap_poly_list(domain = "akcan")
}
\arguments{
\item{domain}{defaults to \code{"akcan"}, the standard SNAP geospatial extent.
The other option is \code{"ak"} pertaining to the classic ALFRESCO model "statewide" extent template.}
}
\value{
a list.
}
\description{
Generate a nested list of spatial polygons data frames from maps in the `snappoly` package.
}
\details{
The list is subset to exclude the `canada` and `lcc` maps that extend outside Alaska.
}
\examples{
\dontrun{snap_poly_list()}
}
