% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapprep.R
\name{snapdef}
\alias{snapdef}
\title{A function that returns a list of SNAP defaults such as path names to data sets and lists of available climate variables and models, etc.}
\usage{
snapdef()
}
\value{
a list.
}
\description{
A function that returns a list of SNAP defaults such as path names to data sets and lists of available climate variables and models, etc.
}
\examples{
snapdef()$ar5gcm
}
