% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clim_ar5.R
\name{split_monthly_files}
\alias{split_monthly_files}
\title{Split grouped monthly files into individual files}
\usage{
split_monthly_files(in_dir = snapdef()$ar5dir_dist_monthly,
  out_dir = snapdef()$ar5dir_dist_monthly_split, mc.cores = 32)
}
\arguments{
\item{in_dir}{input directory, e.g., \code{snapdef()$ar5dir_dist_monthly}.}

\item{out_dir}{output directory, e.g., \code{snapdef()$ar5dir_dist_monthly_split}}

\item{mc.cores}{number of CPUs when processing years in parallel. Defaults to 32 assuming Atlas compute node context.}
}
\description{
Split grouped monthly climate files into individual monthly files with \code{01} through \code{12} suffix in file names.
}
\details{
This function trades fewer files for smaller files and can be preferable for data fetched from AWS in Shiny apps.
}
\examples{
\dontrun{
split_monthly_files()
}
}
