% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_upload}
\alias{aws_upload}
\title{Upload data to Amazon Web Services}
\usage{
aws_upload(in_dir, prefix, bkt = "leonawicz", mc.cores = 32)
}
\arguments{
\item{in_dir}{character, input directory.}

\item{prefix}{character, optional prefix.}

\item{bkt}{S3 bucket name, defaults to the author since this is a SNAP developer package.}

\item{mc.cores}{number of CPUs when processing years in parallel. Defaults to 32 assuming Atlas compute node context.}
}
\description{
Upload prepared SNAP data to Amazon Web Services S3 buckets.
}
\details{
This function uploads SNAP data to S3 buckets for use primarily by SNAP Shiny apps.
The files uploaded includes any recursively listed under \code{in_dir}. The names will include only the base names.
A \code{prefix} can be placed before the file base names. Using \code{/} in the prefix allows for the simulation of folders in the S3 bucket.
Remember that S3 buckets technically have a flat structure.

NOTE: This function will not work unless the AWS keys granting permission to the bucket are loaded into the R session.
}
\examples{
\dontrun{
source("aws_key.R") # SEE DETAILS
aws_upload(snapdef()$ar5dir_dist_monthly, "clim/dist/ar5_2km/monthly")
aws_upload(snapdef()$ar5dir_dist_seasonal, "clim/dist/ar5_2km/seasonal")
}
}
