% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clim_ar5.R
\name{clim_dist_monthly}
\alias{clim_dist_monthly}
\title{Extract climate data and compute distributions}
\usage{
clim_dist_monthly(inputs, in_dir = snapdef()$ar5dir,
  out_dir = snapdef()$ar5dir_dist_monthly, na.rm = TRUE,
  density.args = list(n = 200, adjust = 0.1), sample.size = 10000,
  verbose = TRUE, overwrite = FALSE, move_akcan = TRUE, mc.cores = 32)
}
\arguments{
\item{inputs}{data frame of inputs (one row). See details.}

\item{in_dir}{input directory, e.g., \code{snapdef()$ar5dir}.}

\item{out_dir}{output directory, e.g., \code{snapdef()$ar5dir_dist_monthly}}

\item{na.rm}{logical, remove NAs.}

\item{density.args}{arguments list passed to \code{density}.}

\item{sample.size}{numeric, sample size.}

\item{verbose}{logical, verbose progress.}

\item{overwrite}{logical, overwrite existing files.}

\item{move_akcan}{logical, relocate the \code{AK-CAN} domain from \code{Political Boundaries} subdirectory to top level as its own location group.}

\item{mc.cores}{number of CPUs when processing years in parallel. Defaults to 32 assuming Atlas compute node context.}
}
\value{
invisible, writes files.
}
\description{
Extract climate data and estimate monthly spatial probability distributions.
}
\details{
\code{inputs} generally comes from \link{clim_inputs_table}. \code{clim_dist_monthly} processes data sets referred to by
one row of this data frame at a time. Internally processing uses 32 CPUs on an Atlas compute node. It is expected that the different
data sets in the full \code{inputs} be processed serially. See example call below.
}
\examples{
\dontrun{
purrr::walk(1:nrow(inputs), ~slice(inputs, .x) \%>\% clim_dist_monthly())
}
}
