% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clim_ar5.R
\name{clim_dist_seasonal}
\alias{clim_dist_seasonal}
\title{Compute seasonal climate data spatial distributions}
\usage{
clim_dist_seasonal(i, files, in_dir = snapdef()$ar5dir_dist_monthly,
  out_dir = snapdef()$ar5dir_dist_seasonal, density.args = list(n = 200,
  adjust = 0.1))
}
\arguments{
\item{i}{iterator for files.}

\item{files}{vector of input files.}

\item{in_dir}{input directory.}

\item{out_dir}{output directory.}

\item{density.args}{arguments list passed to \code{density}.}
}
\value{
invisible, writes files.
}
\description{
Compute seasonal climate data spatial probability distributions.
}
\details{
Seasons are DJF, MAM, JJA and SON 3-month averages. A fifth "season" of full annual averages is also included.
For efficiency, this function operates on outputs from \code{clim_dist_monthly}. It does not need to redundantly
access source downscaled geotiffs.
}
\examples{
\dontrun{mclapply(seq_along(files), clim_dist_seasonal,
  files = files, out_dir = out_dir, mc.cores = 32)}
}
