% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clim_ar5.R
\name{clim_dist_seasonal}
\alias{clim_dist_seasonal}
\title{Compute seasonal climate data spatial distributions}
\usage{
clim_dist_seasonal(variable, in_dir = snapdef()$ar5dir_dist_monthly,
  out_dir = snapdef()$ar5dir_dist_seasonal, density.args = list(n = 200,
  adjust = 0.1), mc.cores = 32)
}
\arguments{
\item{variable}{character, optional, to split into smaller file batches. See details.}

\item{in_dir}{input directory, e.g., \code{snapdef()$ar5dir_dist_monthly}.}

\item{out_dir}{output directory, e.g., \code{snapdef()$ar5dir_dist_seasonal}}

\item{density.args}{arguments list passed to \code{density}.}

\item{mc.cores}{number of CPUs when processing years in parallel. Defaults to 32 assuming Atlas compute node context.}
}
\description{
Compute seasonal climate data spatial probability distributions.
}
\details{
Seasons are DJF, MAM, JJA and SON 3-month averages. A fifth "season" of full annual averages is also included.
For efficiency, this function operates on outputs from \code{clim_dist_monthly}. It does not need to redundantly
access source downscaled geotiffs.

Use \code{variable} to optionally specify a climate variable file identifier: \code{"pr"}, \code{"tas"}, \code{"tasmin"} or \code{"tasmax"}.
This will be used for pattern matching when listing files inside \code{in_dir}.
}
\examples{
\dontrun{
clim_dist_seasonal() # all variables
clim_dist_seasonal(variable = "pr") # precipitation
}
}
