% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clim_ar5.R
\name{clim_locs}
\alias{clim_locs}
\title{Under development...}
\usage{
clim_locs(x)
}
\arguments{
\item{x}{a data frame.}
}
\description{
Function under development...
}
\examples{
\dontrun{
clim_locs(x)
}
}
