% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clim_ar5.R
\name{clim_locs}
\alias{clim_locs}
\title{Arrange climate by location and compute seasonals.}
\usage{
clim_locs(in_dir = snapdef()$ar5dir_locs_prep,
  out_dir = snapdef()$ar5dir_locs, overwrite = FALSE, mc.cores = 32)
}
\arguments{
\item{in_dir}{input directory containing files produced by \code{\link{clim_locs_prep}}.}

\item{out_dir}{output directory.}

\item{overwrite}{logical, overwrite existing files.}

\item{mc.cores}{number of CPUs when processing years in parallel. Defaults to 32 assuming Atlas compute node context.}
}
\description{
Arrange extracted monthly climate values for point locations into individual files by location
from preliminary monthly files split by RCP, climate model and climate variable.
Compute seasonal/annual means for temperature variables and seasonal/annual totals for precipitation.
}
\examples{
\dontrun{
clim_locs()
}
}
