% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_clim_ar5.R
\name{clim_stats_ar5}
\alias{clim_stats_ar5}
\title{Compute climate statistics}
\usage{
clim_stats_ar5(files, type = "monthly",
  out_dir = snapdef()$ar5dir_dist_stats[1], mc.cores = 32)
}
\arguments{
\item{files}{input files.}

\item{type}{character, \code{"monthly"} or \code{"seasonal"}.}

\item{out_dir}{output directory, e.g. one of the \code{snapdef()$ar5dir_dist_stats} entries.}

\item{mc.cores}{number of processors.}
}
\value{
invisible, writes files.
}
\description{
Compute climate statistics from spatial probability distributions.
}
\details{
For efficiency, this function operates on outputs from \code{clim_dist_monthly} and \code{clim_dist_seasonal}.
It does not need to redundantly access source downscaled geotiffs. This function is specific to AR5 outputs in the current implementation.
}
\examples{
\dontrun{clim_stats_ar5(files)}
}
