% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_cells.R
\name{save_poly_cells}
\alias{save_poly_cells}
\title{Save rds files containing raster cell tables for polygons}
\usage{
save_poly_cells(file_akcan = "cells_akcan1km2km.rds",
  file_ak = "cells_ak1km.rds", out_dir = snapdef()$celldir,
  akcan1km = snapdef()$template_akcan1km,
  akcan2km = snapdef()$template_akcan2km, ak1km = snapdef()$template_ak1km,
  mc.cores = 32)
}
\arguments{
\item{file_akcan}{file name for cell index table based on the Alaska/western Canada SNAP spatial domain.}

\item{file_ak}{file name for cell index table based on the Alaska "statewide" classic ALFRESCO SNAP spatial domain.}

\item{out_dir}{output directory.}

\item{akcan1km}{Alaska/western Canada domain 1-km ALFRESCO template raster geotiff input file name.}

\item{akcan2km}{Alaska/western Canada domain 2-km climate template raster geotiff input file name.}

\item{ak1km}{Alaska domain 1-km ALFRESCO template raster geotiff input file name.}

\item{mc.cores}{number of processors.}
}
\value{
invisible, writes files.
}
\description{
Save rds files containing data frames of grid cell indices for the intersection of common SNAP spatial polygons
with SNAP template raster layers. Indices are stored for two conditions: matching to all raster cell indices and matching to
new cell indices based on first removing all NA-valued cells and re-indexing.
}
\examples{
\dontrun{save_poly_cells()}
}
